requirements.txt
apturl==0.5.2
asgiref==3.7.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.1
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==41.0.5
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.6
django-allauth==0.57.0
django-cleanup==8.0.0
django-cors-headers==4.3.0
django-filter==23.3
django-rest-auth==0.9.5
django-suit==2.0a2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
httplib2==0.14.0
idna==3.4
importlib-metadata==6.8.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.5
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
oauthlib==3.2.2
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==10.1.0
protobuf==3.6.1
psycopg2-binary==2.9.9
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python3-openid==3.2.0
pytz==2023.3.post1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.31.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
sqlparse==0.4.4
systemd-python==234
typing-extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.7
wadllib==1.3.3
xkit==0.0.0
zipp==3.17.0
